requirements.txt
setuptools==41.0.1
tencentcloud_sdk_python==3.0.65
